requirements.txt
# Python 3.11+
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langsmith>=0.2.0
azure-identity>=1.15.0
openai>=1.0.0

# Document loading and vector store
langchain-community>=0.3.0
chromadb>=0.4.0
pypdf>=4.0.0

# Eval and utilities
python-dotenv>=1.0.0

# Web UI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Optional: OpenTelemetry when LangSmith is not used
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
